Fix partially-dynamic value return bug
Previously, if a value contained a dynamic aspect, the attribute value
returned by Sprig was just the dynamic component of the entire value.

This commit fixes this bug, allowing dynamic components to be returned
within larger strings. At the same time, we still allow the return of
non-string data structures from values that are entirely dynamic.